--- src/main/java/dao/JanggiDao.java
@@ -1,188 +1,0 @@
-package dao;
-
-import config.DBConnection;
-import dto.PieceDto;
-import dto.PositionDto;
-import dto.SavedPieceDto;
-import java.io.IOException;
-import java.io.InputStream;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
-
-public class JanggiDao {
-
-    public void createTable() {
-        try (Connection connection = DBConnection.getConnection();
-             InputStream inputStream = getClass().getClassLoader().getResourceAsStream("init.sql")) {
-            if (inputStream == null) {
-                throw new RuntimeException("init.sql 파일을 찾을 수 없습니다.");
-            }
-
-            String sql = new String(inputStream.readAllBytes());
-            try (PreparedStatement pstmt = connection.prepareStatement(sql)) {
-                pstmt.execute();
-            }
-        } catch (SQLException | IOException e) {
-            System.out.println("테이블 생성 중 에러 발생: " + e.getMessage());
-        }
-    }
-
-    public int createNewGame(String initialTurn) {
-        String sql = "INSERT INTO game_state (turn) VALUES (?)";
-
-        try (Connection connection = DBConnection.getConnection();
-             PreparedStatement pstmt = connection.prepareStatement(sql, PreparedStatement.RETURN_GENERATED_KEYS)) {
-
-            pstmt.setString(1, initialTurn);
-            pstmt.executeUpdate();
-
-            try (ResultSet rs = pstmt.getGeneratedKeys()) {
-                if (rs.next()) {
-                    return rs.getInt(1);
-                }
-            }
-        } catch (SQLException e) {
-            System.out.println("새 게임 생성 중 에러 발생: " + e.getMessage());
-        }
-        throw new RuntimeException("새 게임 생성에 실패했습니다.");
-    }
-
-    public void updateTurn(int gameId, String currentTurn) {
-        String sql = "UPDATE game_state SET turn = ? WHERE id = ?";
-
-        try (Connection connection = DBConnection.getConnection();
-             PreparedStatement pstmt = connection.prepareStatement(sql)) {
-
-            pstmt.setString(1, currentTurn);
-            pstmt.setInt(2, gameId);
-            pstmt.executeUpdate();
-        } catch (SQLException e) {
-            System.out.println("턴 정보 업데이트 중 에러 발생: " + e.getMessage());
-        }
-    }
-
-    public void savePiecePosition(int gameId, PieceDto pieceDto, PositionDto positionDto) {
-        String sql = "INSERT INTO piece_state (game_id, country, piece_name, row_pos, col_pos) VALUES (?, ?, ?, ?, ?)";
-
-        try (Connection connection = DBConnection.getConnection();
-             PreparedStatement pstmt = connection.prepareStatement(sql)) {
-
-            pstmt.setInt(1, gameId);
-            pstmt.setString(2, pieceDto.countryName());
-            pstmt.setString(3, pieceDto.pieceName());
-            pstmt.setInt(4, positionDto.row());
-            pstmt.setInt(5, positionDto.col());
-            pstmt.executeUpdate();
-        } catch (SQLException e) {
-            System.out.println("보드 상태 저장 중 에러 발생: " + e.getMessage());
-        }
-    }
-
-    public void movePiece(int gameId, int beforeRow, int beforeCol, int afterRow, int afterCol) {
-        String sql = "UPDATE piece_state SET row_pos = ?, col_pos = ? WHERE game_id = ? AND row_pos = ? AND col_pos = ?";
-
-        try (Connection connection = DBConnection.getConnection();
-             PreparedStatement pstmt = connection.prepareStatement(sql)) {
-
-            pstmt.setInt(1, afterRow);
-            pstmt.setInt(2, afterCol);
-            pstmt.setInt(3, gameId);
-            pstmt.setInt(4, beforeRow);
-            pstmt.setInt(5, beforeCol);
-            pstmt.executeUpdate();
-        } catch (SQLException e) {
-            System.out.println("말 이동 업데이트 중 에러 발생: " + e.getMessage());
-        }
-    }
-
-    public void deletePiece(int gameId, int row, int col) {
-        String sql = "DELETE FROM piece_state WHERE game_id = ? AND row_pos = ? AND col_pos = ?";
-
-        try (Connection connection = DBConnection.getConnection();
-             PreparedStatement pstmt = connection.prepareStatement(sql)) {
-
-            pstmt.setInt(1, gameId);
-            pstmt.setInt(2, row);
-            pstmt.setInt(3, col);
-            pstmt.executeUpdate();
-        } catch (SQLException e) {
-            System.out.println("말 제거 중 에러 발생: " + e.getMessage());
-        }
-    }
-
-    public List<Integer> getSavedGames() {
-        String sql = "SELECT id FROM game_state";
-        List<Integer> gameId = new ArrayList<>();
-        try (Connection connection = DBConnection.getConnection();
-             PreparedStatement pstmt = connection.prepareStatement(sql);
-             ResultSet rs = pstmt.executeQuery()) {
-
-            while (rs.next()) {
-                gameId.add(rs.getInt("id"));
-            }
-        } catch (SQLException e) {
-            System.out.println("게임 목록 조회 중 에러 발생: " + e.getMessage());
-        }
-
-        return gameId;
-    }
-
-    public List<SavedPieceDto> getSavedBoard(int gameId) {
-        String sql = "SELECT country, piece_name, row_pos, col_pos FROM piece_state WHERE game_id = ?";
-        List<SavedPieceDto> savedPieceDtos = new ArrayList<>();
-        try (Connection connection = DBConnection.getConnection();
-             PreparedStatement pstmt = connection.prepareStatement(sql)) {
-            pstmt.setInt(1, gameId);
-
-            try (ResultSet rs = pstmt.executeQuery()) {
-                while (rs.next()) {
-                    savedPieceDtos.add(
-                            new SavedPieceDto(
-                                    rs.getString("country"),
-                                    rs.getString("piece_name"),
-                                    rs.getInt("row_pos"),
-                                    rs.getInt("col_pos")));
-                }
-            }
-        } catch (SQLException e) {
-            System.out.println("보드 정보 조회 중 에러 발생: " + e.getMessage());
-        }
-
-        return savedPieceDtos;
-    }
-
-    public String getSavedTurn(int gameId) {
-        String sql = "SELECT turn FROM game_state WHERE id = ?";
-        try (Connection connection = DBConnection.getConnection();
-             PreparedStatement pstmt = connection.prepareStatement(sql)) {
-            pstmt.setInt(1, gameId);
-            try (ResultSet rs = pstmt.executeQuery()) {
-                if (rs.next()) {
-                    return rs.getString("turn");
-                }
-            }
-        } catch (SQLException e) {
-            System.out.println("보드 정보 조회 중 에러 발생: " + e.getMessage());
-        }
-
-        return "";
-    }
-
-    public void deleteGame(int gameId) {
-        String sql = "DELETE FROM game_state WHERE id = ?";
-
-        try (Connection connection = DBConnection.getConnection();
-             PreparedStatement pstmt = connection.prepareStatement(sql)) {
-
-            pstmt.setInt(1, gameId);
-            pstmt.executeUpdate();
-
-        } catch (SQLException e) {
-            System.out.println("게임 정보 삭제 중 에러 발생: " + e.getMessage());
-        }
-    }
-}
